Time in navbar is now live.
