+++ frontend/src/utils/timeInTimezone.js
@@ -1,0 +1,13 @@
+// Calculate time for a given timezone
+function timeInTimezone(timezone) {
+    let options = {
+      timeZone: timezone.trim(),
+      hour: 'numeric',
+      minute: 'numeric',
+      hour12: false
+    }
+    const formatter = new Intl.DateTimeFormat([], options);
+    return formatter.format(new Date());
+}
+
+export default timeInTimezone;
